Фикс зависимостей
Добавил ограничения на удаления зелий и ингредиентов если они как то связаны друг с другом или лежат на складе. Так же добавлены мелкие улучшения тут и там и тама

+++ src/main/java/com/potion/ISPotion/Controllers/IngredientController.java
@@ -1,8 +1,10 @@
 package com.potion.ISPotion.Controllers;
 
 import com.potion.ISPotion.Classes.Ingredient;
+import com.potion.ISPotion.Classes.Potion;
 import com.potion.ISPotion.Classes.Role;
-import com.potion.ISPotion.repo.UserRepository;
+import com.potion.ISPotion.Classes.StorageCell;
+import com.potion.ISPotion.repo.*;
 import com.potion.ISPotion.utils.AuthUtils;
 import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,7 +13,6 @@ import org.springframework.security.core.annotation.CurrentSecurityContext;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
-import com.potion.ISPotion.repo.IngredientRepository;
 
 import java.util.*;
 
@@ -22,6 +23,10 @@ public class IngredientController {
 
     @Autowired
     private UserRepository userRepository;
+    @Autowired
+    private PotionRepository potionRepository;
+    @Autowired
+    private StorageCellRepository storageCellRepository;
 
     @ModelAttribute("permissionsParts")
     public Set<String> headerPermission(@CurrentSecurityContext(expression="authentication")
@@ -71,6 +76,22 @@ public class IngredientController {
             return "redirect:/home";
 
         Iterable<Ingredient> ingredients = ingredientRepository.findAll();
+        ArrayList<Long> notDeleteableIds = new ArrayList<>();
+        for (Ingredient ingr: ingredients
+        ) {
+            Set<Potion> potionsWithIngredient = potionRepository.findAllByIngredients(ingr);
+            Set<StorageCell> storagedIngredients = storageCellRepository.findAllByEntityId(ingr.getId());
+            if(!potionsWithIngredient.isEmpty() || !storagedIngredients.isEmpty()) {
+                notDeleteableIds.add(ingr.getId());
+            }
+        }
+
+        if(!notDeleteableIds.isEmpty()) {
+            String errorMessage = "Вы не можете удалить этот объект, потому что существуют взаимосвязи с другими сущностями";
+            model.addAttribute("errorMessage", errorMessage);
+        }
+
+        model.addAttribute("notDeleteableIds", notDeleteableIds);
         model.addAttribute("ingredients", ingredients );
         model.addAttribute("title", "Ингредиенты");
         return "ingredient";
@@ -95,6 +116,7 @@ public class IngredientController {
             return "redirect:/ingredient";
         }
         Ingredient ingredient = ingredientRepository.findById(id).orElseThrow();
+
         model.addAttribute("ingredient", ingredient );
         model.addAttribute("title", "Ингредиенты");
         return "ingredient-edit";
@@ -145,7 +167,7 @@ public class IngredientController {
     }
     @PostMapping("/ingredient/delete/{id}")
     public String ingredientDelete(@CurrentSecurityContext(expression="authentication")
-                                       Authentication authentication, @PathVariable(value = "id") long id) {
+                                       Authentication authentication, @PathVariable(value = "id") long id, Model model) {
         Collection<Role> allowedRoles = new HashSet<>(Arrays.asList(
                 Role.DIRECTOR,
                 Role.ADMIN
@@ -161,6 +183,13 @@ public class IngredientController {
         if (!ingredientRepository.existsById(id)) {
             return "redirect:/ingredient";
         }
+
+        Ingredient ingredient = ingredientRepository.findById(id).orElseThrow();
+        Set<Potion> potionsWithIngredient = potionRepository.findAllByIngredients(ingredient);
+        if(!potionsWithIngredient.isEmpty()) {
+            return "redirect:/ingredient";
+        }
+
         ingredientRepository.deleteById(id);
         return "redirect:/ingredient";
     }
